refactor(python): separate auth.base transport
