+++ crates/runner/mitm-addon/tests/conftest.py
@@ -28,6 +28,7 @@ from mitmproxy.test import tflow, tutils
 import anthropic_accounting
 import auth
 import auth_base_forwarder
+import auth_base_transport
 import aws_sigv4_body_admission
 import builtin_connector_diagnostics
 import claude_output_timing
@@ -60,6 +61,7 @@ def _reset_module_state() -> Iterator[None]:
     reset them around each case to avoid cross-test callbacks.
     """
     auth_base_forwarder.reset_forward_request_state_for_tests()
+    auth_base_transport.reset_transport_state_for_tests()
     firewall_auth_client.reset_transport_state_for_tests()
     aws_sigv4_body_admission.reset_for_tests()
     builtin_connector_diagnostics.reset_cache_for_tests()
@@ -87,6 +89,7 @@ def _reset_module_state() -> Iterator[None]:
     codex_output_timing.reset_for_tests()
     logging_utils.reset_log_writer_for_tests()
     auth_base_forwarder.reset_forward_request_state_for_tests()
+    auth_base_transport.reset_transport_state_for_tests()
     firewall_auth_client.reset_transport_state_for_tests()
     aws_sigv4_body_admission.reset_for_tests()
     builtin_connector_diagnostics.reset_cache_for_tests()
